user-service/requirements.txt
django 
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
python-decouple==3.8
django-allauth
social-auth-app-django
dj-rest-auth
drf-spectacular